broke out module component and hooked up domainlist
