+++ retrotide/src/containers/DomainSearch.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../components/Button';
+import ModuleBuilder from '../components/ModuleBuilder'
 
 class DomainSearch extends React.Component {
 
@@ -20,105 +20,17 @@ class DomainSearch extends React.Component {
     }
   }
 
-  getAllDomains = () => {
-    let allDomains = [];
-    for (var DomainObject in this.state.PKSDomainList) {
-      allDomains.push(this.state.PKSDomainList[DomainObject]);
-    }
-    return allDomains;
-  };
-
-  getPresentDomains = () => {
-    let presentDomains = [];
-    for (var DomainObject in this.state.PKSDomainList) {
-      if (this.state.PKSDomainList[DomainObject].present) {
-        presentDomains.push(this.state.PKSDomainList[DomainObject]);
-      }
-    }
-    return presentDomains;
-  };
-
-  // getPresentDomainsInOrder
-
-  selectModuleType = e => {
-    let value = e.target.value;
-    if(value === 'loading') {
-      // set AT, ACP, KS buttons true
-      // preload ACP in sandbox
-    } else if(value === 'reducing') {
-      // set all but TE buttons true
-      // preload KS AT ACP in sandbox
-    } else if(value === 'terminating') {
-      // set all and TE buttons true
-      // preload KS ACP TE in sandbox
-    } else {
-      // what did you do
-    }
-  }
-
-  insertDomain = Domain => {
-    let selectedDomain = this.state.PKSDomainList[Domain];
-
-    if(selectedDomain.present) {
-      console.log("ERR that domain is already selected " + selectedDomain.domainName);
-      return -1;
-    } else {
-      // we'll need some logic here to add multiple nodes depending on selected name
-      let insertDomain = {
-        domainName: Domain,
-        present: true,
-      }
-      let updatedDomainList = {
-        ...this.state.PKSDomainList,
-        [Domain]: insertDomain,
-      };
-      this.setState({PKSDomainList: updatedDomainList});      
-    }
-  }
-
-  deleteDomain = Domain => {
-    let deleteDomain = {
-      domainName: Domain,
-      present: false,
-    }
-    let updatedDomainList = {
-      ...this.state.PKSDomainList,
-      [Domain]: deleteDomain,
-    };
-    this.setState({PKSDomainList: updatedDomainList});
-  }
+  // select PKS or NRBS
+  // list of all modulebuilders
+  // add module button
+  // delete module button
+  // submit button
 
   render() {
     return (
       <div className='DomainSearch'>
-        <h3>Construct Module</h3> 
-        <div className="DomainToolbox">
-          <div className="DomainButtonList">
-            <select className="ModuleType" onChange={ (e) => {this.selectModuleType(e)} }>
-              <option selected value="loading">Loading</option>
-              <option value="reducing">Reducing</option>
-              <option value="terminating">Terminating</option>
-            </select>
-            {this.getAllDomains().map((DomainButton, index) => (
-              <Button className='addDomainButton' key={index} onClick={ () => {this.insertDomain(DomainButton.domainName)} }>
-                {DomainButton.domainName}
-              </Button>
-              ))
-            }
-          </div>
-          <div className="DomainSandbox">
-            {this.getPresentDomains().map((DomainDiv, index) => (
-                <div key={DomainDiv.domainName + index}>
-                  <div className={"Domain " + DomainDiv.domainName}>
-                    {DomainDiv.domainName}
-                    <div className="handle"> -> </div>
-                    <div className="deleteIcon" onClick={ () => {this.deleteDomain(DomainDiv.domainName)} }> X </div>
-                  </div>
-                </div>
-              ))
-            }
-          </div>
-        </div>
+        <h3>Construct Modules</h3> 
+        <ModuleBuilder domainList={this.state.PKSDomainList} />
       </div>
     )
   }
